MC-40 - Created PasswordInputComponents
